main.py
import os
import sys

sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))

from dotenv import load_dotenv
load_dotenv()

from ui.app import create_app

if __name__ == "__main__":
    demo = create_app()
    demo.launch(server_port=7860, share=False)
