requirements.txt
langchain>=0.2.0
langchain-core>=0.2.0
langchain-google-genai>=1.0.0
langchain-community>=0.2.0
langgraph>=0.1.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
pypdf>=4.0.0
python-docx>=1.1.0
tavily-python>=0.3.0
gradio>=4.40.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
langchain-huggingface>=0.0.3
